Extract Currency struct into it's own file
