+++ Sources/Afterpay/Model/Configuration.swift
@@ -8,25 +8,6 @@
 
 import Foundation
 
-private let posixLocale = Locale(identifier: "en_US_POSIX")
-private let unknownCurrencyName = "Unknown Currency"
-
-struct Currency {
-
-  let code: String
-
-  init?(currencyCode: String) {
-    let currencyName = posixLocale.localizedString(forCurrencyCode: currencyCode)
-
-    guard currencyName != nil, currencyName != unknownCurrencyName else {
-      return nil
-    }
-
-    self.code = currencyCode
-  }
-
-}
-
 /// Describes errors encountered when constructing a `Configuration`
 public enum ConfigurationError: LocalizedError, Equatable {
